wildberries_task/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
yarl==1.8.1
